# backend/internal/interfaces/handler/http/handler/message_handler.go
package handler

import (
	"net/http"
	"time"

	"github.com/labstack/echo/v4"
	openapi_types "github.com/oapi-codegen/runtime/types"

	"github.com/newt239/chat/internal/infrastructure/utils"
	openapi "github.com/newt239/chat/internal/openapi_gen"
	messageuc "github.com/newt239/chat/internal/usecase/message"
)

type MessageHandler struct {
	MessageUC messageuc.MessageUseCase
}

func (h *MessageHandler) ListMessagesWithThread(c echo.Context, channelId openapi_types.UUID, params openapi.ListMessagesWithThreadParams) error {
	userID, ok := c.Get("userID").(string)
	if !ok {
		return utils.HandleAuthError()
	}

	limit := 20
	if params.Limit != nil {
		limit = *params.Limit
	}

	var sinceTime *time.Time
	if params.Since != nil {
		sinceTime = params.Since
	}

	var untilTime *time.Time
	if params.Until != nil {
		untilTime = params.Until
	}

	input := messageuc.ListMessagesInput{
		ChannelID: channelId.String(),
		UserID:    userID,
		Limit:     limit,
		Since:     sinceTime,
		Until:     untilTime,
	}

	listRes, err := h.MessageUC.ListMessages(c.Request().Context(), input)
	if err != nil {
		return handleUseCaseError(err)
	}

	outputs, err := h.MessageUC.ListMessagesWithThread(c.Request().Context(), input)
	if err != nil {
		return handleUseCaseError(err)
	}

	return c.JSON(http.StatusOK, map[string]any{
		"messages": outputs,
		"hasMore":  listRes.HasMore,
	})
}

func (h *MessageHandler) GetThreadReplies(c echo.Context, messageId openapi_types.UUID, params openapi.GetThreadRepliesParams) error {
	userID, ok := c.Get("userID").(string)
	if !ok {
		return utils.HandleAuthError()
	}

	input := messageuc.GetThreadRepliesInput{
		MessageID: messageId.String(),
		UserID:    userID,
	}

	output, err := h.MessageUC.GetThreadReplies(c.Request().Context(), input)
	if err != nil {
		switch err {
		case messageuc.ErrParentMessageNotFound:
			return echo.NewHTTPError(http.StatusNotFound, err.Error())
		case messageuc.ErrUnauthorized:
			return echo.NewHTTPError(http.StatusForbidden, err.Error())
		default:
			return handleUseCaseError(err)
		}
	}

	return c.JSON(http.StatusOK, output)
}

func (h *MessageHandler) GetThreadMetadata(c echo.Context, messageId openapi_types.UUID) error {
	userID, ok := c.Get("userID").(string)
	if !ok {
		return utils.HandleAuthError()
	}

	input := messageuc.GetThreadMetadataInput{
		MessageID: messageId.String(),
		UserID:    userID,
	}

	output, err := h.MessageUC.GetThreadMetadata(c.Request().Context(), input)
	if err != nil {
		switch err {
		case messageuc.ErrParentMessageNotFound:
			return echo.NewHTTPError(http.StatusNotFound, err.Error())
		case messageuc.ErrUnauthorized:
			return echo.NewHTTPError(http.StatusForbidden, err.Error())
		default:
			return handleUseCaseError(err)
		}
	}

	return c.JSON(http.StatusOK, output)
}

func (h *MessageHandler) ListMessages(c echo.Context, channelId openapi_types.UUID, params openapi.ListMessagesParams) error {
	userID, ok := c.Get("userID").(string)
	if !ok {
		return utils.HandleAuthError()
	}

	limit := 20
	if params.Limit != nil {
		limit = *params.Limit
	}

	var sinceTime *time.Time
	if params.Since != nil {
		sinceTime = params.Since
	}

	var untilTime *time.Time
	if params.Until != nil {
		untilTime = params.Until
	}

	input := messageuc.ListMessagesInput{
		ChannelID: channelId.String(),
		UserID:    userID,
		Limit:     limit,
		Since:     sinceTime,
		Until:     untilTime,
	}

	messages, err := h.MessageUC.ListMessages(c.Request().Context(), input)
	if err != nil {
		return handleUseCaseError(err)
	}

	return c.JSON(http.StatusOK, messages)
}

func (h *MessageHandler) CreateMessage(c echo.Context, channelId openapi_types.UUID) error {
	userID, ok := c.Get("userID").(string)
	if !ok {
		return utils.HandleAuthError()
	}

	var req openapi.CreateMessageRequest
	if err := c.Bind(&req); err != nil {
		return utils.HandleBindError(err)
	}

	if err := c.Validate(&req); err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, err.Error())
	}

	var parentID *string
	if req.ParentId != nil {
		parentIDStr := req.ParentId.String()
		parentID = &parentIDStr
	}

	var attachmentIDs []string
	if req.AttachmentIds != nil {
		attachmentIDs = make([]string, len(*req.AttachmentIds))
		for i, id := range *req.AttachmentIds {
			attachmentIDs[i] = id.String()
		}
	}

	input := messageuc.CreateMessageInput{
		ChannelID:     channelId.String(),
		UserID:        userID,
		Body:          req.Body,
		ParentID:      parentID,
		AttachmentIDs: attachmentIDs,
	}

	message, err := h.MessageUC.CreateMessage(c.Request().Context(), input)
	if err != nil {
		return handleUseCaseError(err)
	}

	return c.JSON(http.StatusCreated, message)
}

func (h *MessageHandler) UpdateMessage(c echo.Context, messageId openapi_types.UUID) error {
	userID, ok := c.Get("userID").(string)
	if !ok {
		return utils.HandleAuthError()
	}

	var req openapi.UpdateMessageRequest
	if err := c.Bind(&req); err != nil {
		return utils.HandleBindError(err)
	}

	if err := c.Validate(&req); err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, err.Error())
	}

	input := messageuc.UpdateMessageInput{
		MessageID: messageId.String(),
		EditorID:  userID,
		Body:      req.Body,
	}

	message, err := h.MessageUC.UpdateMessage(c.Request().Context(), input)
	if err != nil {
		return mapMessageError(err)
	}

	return c.JSON(http.StatusOK, message)
}

func (h *MessageHandler) DeleteMessage(c echo.Context, messageId openapi_types.UUID) error {
	userID, ok := c.Get("userID").(string)
	if !ok {
		return utils.HandleAuthError()
	}

	input := messageuc.DeleteMessageInput{
		MessageID:  messageId.String(),
		ExecutorID: userID,
	}

	if err := h.MessageUC.DeleteMessage(c.Request().Context(), input); err != nil {
		return mapMessageError(err)
	}

	return c.NoContent(http.StatusNoContent)
}

func mapMessageError(err error) error {
	switch err {
	case messageuc.ErrMessageNotFound, messageuc.ErrChannelNotFound:
		return echo.NewHTTPError(http.StatusNotFound, err.Error())
	case messageuc.ErrUnauthorized:
		return echo.NewHTTPError(http.StatusForbidden, err.Error())
	case messageuc.ErrMessageAlreadyDeleted, messageuc.ErrCannotEditDeleted:
		return echo.NewHTTPError(http.StatusBadRequest, err.Error())
	default:
		return handleUseCaseError(err)
	}
}