# openapi/bundled.yaml
openapi: 3.1.0
info:
  title: Chat API
  version: 0.1.0
  description: Slack-like communication application API
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/channels/{channelId}/messages:
    get:
      operationId: listMessages
      summary: List messages in channel
      security:
        - bearerAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: since
          in: query
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: createMessage
      summary: Create a new message
      security:
        - bearerAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
      responses:
        '201':
          description: Message created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/channels/{channelId}/messages/with-threads:
    get:
      operationId: listMessagesWithThread
      summary: List messages in channel with thread metadata
      security:
        - bearerAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: since
          in: query
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of messages with thread metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Message'
                        - type: object
                          properties:
                            threadMetadata:
                              $ref: '#/components/schemas/ThreadMetadata'
                              nullable: true
                  hasMore:
                    type: boolean
                required:
                  - messages
                  - hasMore
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/channels/{channelId}/reads:
    post:
      operationId: updateReadState
      summary: Update channel read state
      security:
        - bearerAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReadStateRequest'
      responses:
        '200':
          description: Read state updated
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/channels/{channelId}/unread_count:
    get:
      operationId: getUnreadCount
      summary: Get unread message count
      security:
        - bearerAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Unread count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnreadCountResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/messages/{messageId}:
    patch:
      operationId: updateMessage
      summary: Update a message
      description: Updates a message body (author or admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMessageRequest'
      responses:
        '200':
          description: Message updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Bad request or message already deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not message author or admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteMessage
      summary: Delete a message
      description: Deletes a message (author or admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Message deleted successfully
        '400':
          description: Bad request or message already deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not message author or admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/messages/{messageId}/reactions:
    get:
      operationId: listReactions
      summary: List reactions for a message
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of reactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListReactionsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: addReaction
      summary: Add a reaction to a message
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddReactionRequest'
      responses:
        '201':
          description: Reaction added
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/messages/{messageId}/reactions/{emoji}:
    delete:
      operationId: removeReaction
      summary: Remove a reaction from a message
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: emoji
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reaction removed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Reaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/messages/{messageId}/thread:
    get:
      operationId: getThreadReplies
      summary: Get thread replies for a message
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Thread replies retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadRepliesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/messages/{messageId}/thread/metadata:
    get:
      operationId: getThreadMetadata
      summary: Get thread metadata for a message
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Thread metadata retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadMetadata'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Message or metadata not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AddReactionRequest:
      type: object
      properties:
        emoji:
          type: string
          minLength: 1
      required:
        - emoji
    Attachment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        messageId:
          type: string
          format: uuid
        fileName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - messageId
        - fileName
        - mimeType
        - sizeBytes
        - createdAt
...
    CreateMessageRequest:
      type: object
      properties:
        body:
          type: string
          minLength: 1
        parentId:
          type: string
          format: uuid
        attachmentIds:
          type: array
          items:
            type: string
            format: uuid
      required:
        - body
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ListReactionsResponse:
      type: object
      properties:
        reactions:
          type: array
          items:
            $ref: '#/components/schemas/ReactionWithUser'
      required:
        - reactions
...
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        channelId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
          nullable: true
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
          nullable: true
        deletedAt:
          type: string
          format: date-time
          nullable: true
        isDeleted:
          type: boolean
        deletedBy:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            displayName:
              type: string
            avatarUrl:
              type: string
              nullable: true
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
      required:
        - id
        - channelId
        - userId
        - body
        - createdAt
        - isDeleted
...
    MessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        hasMore:
          type: boolean
      required:
        - messages
        - hasMore
...
    ReactionWithUser:
      type: object
      properties:
        messageId:
          type: string
          format: uuid
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            displayName:
              type: string
            avatarUrl:
              type: string
              nullable: true
          required:
            - id
            - displayName
        emoji:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - messageId
        - user
        - emoji
        - createdAt
...
    ThreadMetadata:
      type: object
      properties:
        messageId:
          type: string
          format: uuid
        replyCount:
          type: integer
          minimum: 0
        lastReplyAt:
          type: string
          format: date-time
          nullable: true
        lastReplyUser:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            displayName:
              type: string
            avatarUrl:
              type: string
              nullable: true
          required:
            - id
            - displayName
        participantUserIds:
          type: array
          items:
            type: string
            format: uuid
      required:
        - messageId
        - replyCount
        - participantUserIds
    ThreadRepliesResponse:
      type: object
      properties:
        parentMessage:
          $ref: '#/components/schemas/Message'
        replies:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        hasMore:
          type: boolean
      required:
        - parentMessage
        - replies
        - hasMore
    UnreadCountResponse:
      type: object
      properties:
        count:
          type: integer
          minimum: 0
      required:
        - count
...
    UpdateMessageRequest:
      type: object
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 10000
      required:
        - body
    UpdateReadStateRequest:
      type: object
      properties:
        lastReadAt:
          type: string
          format: date-time
      required:
        - lastReadAt
...